automatically backed up by learn

+++ lib/create.sql
@@ -8,14 +8,14 @@ CREATE TABLE projects (
   id INTEGER PRIMARY KEY,
   title TEXT,
   category TEXT,
-  funding_goal FLOAT,
+  funding_goal INTEGER,
   start_date DATE,
   end_date DATE
 );
 
 CREATE TABLE pledges (
   id INTEGER PRIMARY KEY,
-  amount FLOAT,
+  amount INTEGER,
   user_id INTEGER,
   project_id INTEGER
 );
